+++ packages/react-ui/components/Autocomplete/Autocomplete.tsx
@@ -25,6 +25,7 @@ import { getDOMRect } from '../../lib/dom/getDOMRect';
 
 import { styles } from './Autocomplete.styles';
 import { AutocompleteLocale, AutocompleteLocaleHelper } from './locale';
+import { getAutocompleteTheme } from './getAutocompleteTheme';
 
 function match(pattern: string, items: string[]) {
   if (!pattern || !items) {
@@ -189,11 +190,13 @@ export class Autocomplete extends React.Component<AutocompleteProps, Autocomplet
     return (
       <ThemeContext.Consumer>
         {(theme) => {
-          this.theme = theme;
+          this.theme = getAutocompleteTheme(theme);
           return (
-            <CommonWrapper rootNodeRef={this.setRootNode} {...this.props}>
-              {this.renderMain}
-            </CommonWrapper>
+            <ThemeContext.Provider value={this.theme}>
+              <CommonWrapper rootNodeRef={this.setRootNode} {...this.props}>
+                {this.renderMain}
+              </CommonWrapper>
+            </ThemeContext.Provider>
           );
         }}
       </ThemeContext.Consumer>
@@ -279,13 +282,7 @@ export class Autocomplete extends React.Component<AutocompleteProps, Autocomplet
     }
 
     return (
-      <DropdownContainer
-        offsetY={1}
-        getParent={this.getAnchor}
-        align={menuAlign}
-        disablePortal={disablePortal}
-        menuPos={menuPos}
-      >
+      <DropdownContainer getParent={this.getAnchor} align={menuAlign} disablePortal={disablePortal} menuPos={menuPos}>
         <Menu {...menuProps}>{this.getItems()}</Menu>
       </DropdownContainer>
     );