+++ packages/react-ui/components/Input/Input.tsx
@@ -1,40 +1,54 @@
 // TODO: Enable this rule in functional components.
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import invariant from 'invariant';
-import React, { AriaAttributes } from 'react';
+import React, { AriaAttributes, HTMLAttributes } from 'react';
 import raf from 'raf';
+import warning from 'warning';
 
-import { isIE11, isEdge } from '../../lib/client';
+import { isEdge, isIE11 } from '../../lib/client';
 import { isKeyBackspace, isKeyDelete, someKeys } from '../../lib/events/keyboard/identifiers';
 import { needsPolyfillPlaceholder } from '../../lib/needsPolyfillPlaceholder';
 import { Nullable, Override } from '../../typings/utility-types';
 import { MaskedInput } from '../../internal/MaskedInput';
 import { ThemeContext } from '../../lib/theming/ThemeContext';
 import { Theme } from '../../lib/theming/Theme';
-import { CommonWrapper, CommonProps, CommonWrapperRestProps } from '../../internal/CommonWrapper';
+import { CommonProps, CommonWrapper, CommonWrapperRestProps } from '../../internal/CommonWrapper';
 import { cx } from '../../lib/theming/Emotion';
 import { rootNode, TSetRootNode } from '../../lib/rootNode';
 import { createPropsGetter } from '../../lib/createPropsGetter';
+import { isTheme2022 } from '../../lib/theming/ThemeHelpers';
 
 import { styles } from './Input.styles';
+import { InputLayout } from './InputLayout/InputLayout';
+import { PolyfillPlaceholder } from './InputLayout/PolyfillPlaceholder';
+
+export const inputTypes = ['password', 'text', 'number', 'tel', 'search', 'time', 'date', 'url', 'email'] as const;
 
 export type InputSize = 'small' | 'medium' | 'large';
 export type InputAlign = 'left' | 'center' | 'right';
-export type InputType =
-  | 'password'
-  | 'text'
-  | 'number'
-  | 'tel'
-  | 'search'
-  | 'time'
-  | 'date'
-  | 'url'
-  | 'email'
-  | 'hidden';
+export type InputType = typeof inputTypes[number];
 export type InputIconType = React.ReactNode | (() => React.ReactNode);
+
+export const selectionAllowedTypes: InputType[] = ['text', 'password', 'tel', 'search', 'url'];
+export const selectionErrorMessage = (type: InputType, allowedTypes: InputType[] = selectionAllowedTypes) => {
+  return `<Input />. Selection is not supported by the type "${type}". Types that support selection: ${allowedTypes
+    .map((i) => `"${i}"`)
+    .join(', ')}. Reason: https://developer.mozilla.org/en-US/docs/Web/API/HTMLInputElement/setSelectionRange.`;
+};
+
+export const maskForbiddenTypes: InputType[] = ['number', 'date', 'time'];
+export const maskAllowedTypes: InputType[] = inputTypes.filter((type) => {
+  return !maskForbiddenTypes.includes(type);
+});
+export const maskErrorMessage = (type: InputType, allowedTypes: InputType[] = maskAllowedTypes) => {
+  return `<Input />. Prop "mask" does not support type "${type}". Supported types: ${allowedTypes
+    .map((i) => `"${i}"`)
+    .join(', ')}.`;
+};
 
 export interface InputProps
   extends CommonProps,
+    Pick<HTMLAttributes<unknown>, 'role'>,
     Override<
       React.InputHTMLAttributes<HTMLInputElement>,
       {
@@ -84,7 +98,7 @@ export interface InputProps
         /** Вызывается на label */
         onMouseOver?: React.MouseEventHandler<HTMLLabelElement>;
         /**
-         * Тип. Возможные значения: 'password' | 'text' | 'number' | 'tel' | 'search' | 'time' | 'date' | 'url' | 'email' | 'hidden'
+         * Тип. Возможные значения: 'password' | 'text' | 'number' | 'tel' | 'search' | 'time' | 'date' | 'url' | 'email'
          * */
         type?: InputType;
         /** Значение */
@@ -113,6 +127,13 @@ export interface InputProps
          * @param value значение инпута.
          */
         onUnexpectedInput?: (value: string) => void;
+        /** @ignore */
+        corners?: Partial<
+          Pick<
+            React.CSSProperties,
+            'borderTopRightRadius' | 'borderBottomRightRadius' | 'borderBottomLeftRadius' | 'borderTopLeftRadius'
+          >
+        >;
         /**
          * Атрибут для указания id элемента(-ов), описывающих его
          */
@@ -130,7 +151,7 @@ export const InputDataTids = {
   root: 'Input__root',
 } as const;
 
-type DefaultProps = Required<Pick<InputProps, 'size'>>;
+type DefaultProps = Required<Pick<InputProps, 'size' | 'type'>>;
 
 /**
  * Интерфейс пропсов наследуется от `React.InputHTMLAttributes<HTMLInputElement>`.
@@ -142,6 +163,7 @@ export class Input extends React.Component<InputProps, InputState> {
 
   public static defaultProps: DefaultProps = {
     size: 'small',
+    type: 'text',
   };
 
   private getProps = createPropsGetter(Input.defaultProps);
@@ -157,6 +179,20 @@ export class Input extends React.Component<InputProps, InputState> {
   private blinkTimeout = 0;
   private input: HTMLInputElement | null = null;
   private setRootNode!: TSetRootNode;
+
+  private outputMaskError() {
+    warning(!(this.props.mask && this.canBeUsedWithMask), maskErrorMessage(this.getProps().type));
+  }
+
+  public componentDidMount() {
+    this.outputMaskError();
+  }
+
+  public componentDidUpdate(prevProps: Readonly<InputProps>) {
+    if (this.props.type !== prevProps.type || this.props.mask !== prevProps.mask) {
+      this.outputMaskError();
+    }
+  }
 
   public componentWillUnmount() {
     if (this.blinkTimeout) {
@@ -212,6 +248,14 @@ export class Input extends React.Component<InputProps, InputState> {
    * @param {number} end
    */
   public setSelectionRange(start: number, end: number) {
+    // https://github.com/facebook/react/issues/7769
+    // https://developer.mozilla.org/en-US/docs/Web/API/HTMLInputElement/setSelectionRange
+    if (!selectionAllowedTypes.includes(this.getProps().type)) {
+      warning(false, selectionErrorMessage(this.getProps().type));
+
+      return;
+    }
+
     if (!this.input) {
       throw new Error('Cannot call "setSelectionRange" on unmounted Input');
     }
@@ -247,6 +291,10 @@ export class Input extends React.Component<InputProps, InputState> {
         }}
       </ThemeContext.Consumer>
     );
+  }
+
+  private get canBeUsedWithMask() {
+    return maskForbiddenTypes.includes(this.getProps().type);
   }
 
   /**
@@ -293,13 +341,14 @@ export class Input extends React.Component<InputProps, InputState> {
       onValueChange,
       width,
       error,
+      role,
       warning,
       leftIcon,
       rightIcon,
       borderless,
       value,
       align,
-      type = 'text',
+      type,
       mask,
       maskChar,
       alwaysShowMask,
@@ -311,6 +360,7 @@ export class Input extends React.Component<InputProps, InputState> {
       prefix,
       suffix,
       formatChars,
+      corners,
       'aria-describedby': ariaDescribedby,
       ...rest
     } = props;
@@ -319,7 +369,8 @@ export class Input extends React.Component<InputProps, InputState> {
 
     const labelProps = {
       className: cx(styles.root(this.theme), this.getSizeClassName(), {
-        [styles.focus(this.theme)]: focused,
+        [styles.focus(this.theme)]: focused && !warning && !error,
+        [styles.hovering(this.theme)]: !focused && !disabled && !warning && !error && !borderless,
         [styles.blink(this.theme)]: blinking,
         [styles.borderless()]: borderless && !focused,
         [styles.disabled(this.theme)]: disabled,
@@ -329,7 +380,7 @@ export class Input extends React.Component<InputProps, InputState> {
         [styles.warningFallback(this.theme)]: warning && (isIE11 || isEdge),
         [styles.errorFallback(this.theme)]: error && (isIE11 || isEdge),
       }),
-      style: { width },
+      style: { width, ...corners },
       onMouseEnter,
       onMouseLeave,
       onMouseOver,
@@ -342,6 +393,7 @@ export class Input extends React.Component<InputProps, InputState> {
         [styles.inputDisabled(this.theme)]: disabled,
       }),
       value,
+      role,
       onChange: this.handleChange,
       onFocus: this.handleFocus,
       onKeyDown: this.handleKeyDown,
@@ -355,14 +407,32 @@ export class Input extends React.Component<InputProps, InputState> {
       'aria-describedby': ariaDescribedby,
     };
 
-    const typesDisallowedWithMask: InputType[] = ['number', 'date', 'time', 'hidden'];
     const input =
-      mask && !typesDisallowedWithMask.includes(type)
+      mask && !this.canBeUsedWithMask
         ? this.renderMaskedInput(inputProps, mask)
         : React.createElement('input', inputProps);
 
-    if (type === 'hidden') {
-      return input;
+    if (isTheme2022(this.theme)) {
+      return (
+        <InputLayout
+          leftIcon={leftIcon}
+          rightIcon={rightIcon}
+          prefix={prefix}
+          suffix={suffix}
+          labelProps={labelProps}
+          context={{ disabled: Boolean(disabled), focused, size }}
+        >
+          {input}
+          <PolyfillPlaceholder
+            isMaskVisible={this.isMaskVisible}
+            value={value}
+            defaultValue={this.props.defaultValue}
+            align={align}
+          >
+            {placeholder}
+          </PolyfillPlaceholder>
+        </InputLayout>
+      );
     }
 
     return (
@@ -522,13 +592,6 @@ export class Input extends React.Component<InputProps, InputState> {
     });
 
     if (this.props.selectAllOnFocus) {
-      // https://github.com/facebook/react/issues/7769
-      // https://developer.mozilla.org/en-US/docs/Web/API/HTMLInputElement/setSelectionRange
-      const allowedTypes: InputType[] = ['text', 'password', 'tel', 'search', 'url'];
-      const canBeSelected = !this.props.type || (this.props.type && allowedTypes.includes(this.props.type));
-      if (!canBeSelected) {
-        return;
-      }
       this.input && !isIE11 ? this.selectAll() : this.delaySelectAll();
     }